Path to api_v3 cache is wrong, also, check these dirs actually do exist before attempting to remove to avoid errors in the cronlog.

+++ alpha/crond/kaltura/clear_cache.sh
@@ -24,14 +24,14 @@ echo $$ > $LCK_FILE
 
 
 
-
-echo "`date +%s` start clean v3 `date`" >> $LOG_DIR/clear_cache.log
-#nice -n 19 find /tmp/cache_v3-600 -type f -mmin +1440 -name "cache*" -delete
-/usr/bin/ionice -c3 find $APP_DIR/cache/response/cache_v3-600 -type f -mmin +1440 -name "cache*" -delete
-echo "`date +%s` end clean v3 `date`" >> $LOG_DIR/clear_cache.log
-#nice -n 19 find /tmp/cache_v2 -type f -mmin +1440 -name "cache*" -delete
-/usr/bin/ionice -c3 find /tmp/cache_v2 -type f -mmin +1440 -name "cache*" -delete
+if [ -d "/tmp/cache_v3-600" ];then
+	echo "`date +%s` start clean v3 `date`" >> $LOG_DIR/clear_cache.log
+	/usr/bin/ionice -c3 find /tmp/cache_v3-600 -type f -mmin +1440 -name "cache*" -delete
+fi
+if [ -d "/tmp/cache_v2" ];then
 echo "`date +%s` end clean v2 `date`" >> $LOG_DIR/clear_cache.log
-#nice -n 19 find /tmp -maxdepth 0 -type f -mmin +1440 -name "php*" -delete
+	/usr/bin/ionice -c3 find /tmp/cache_v2 -type f -mmin +1440 -name "cache*" -delete
+fi
+echo "`date +%s` end clean v2 `date`" >> $LOG_DIR/clear_cache.log
 /usr/bin/ionice -c3 find /tmp -maxdepth 1 -type f -mmin +1440 -name "php*" -delete
 echo "`date +%s` end clean php `date`" >> $LOG_DIR/clear_cache.log
